fix(marketing): preserve landing semantics
